requirements.txt
asgiref >= 3.3.4
Django >= 3.2.6
pytz >= 2021.1
sqlparse >= 0.4.1
gunicorn >= 20.1.0
django-heroku >= 0.3.1
django-compressor >= 2.4
django-libsass >= 0.7
djangorestframework >= 3.12.4
django-filter>=2.4.0
Pillow >= 8.2.0